Add key expiration logic
This is possibly overly conservative and paranoid by a couple orders of
magnitude.

Intentionally do not use SERVACK because they're mildly annoying to get at and
aren't authenticated. When we receive a notice authenticated with a key, we
know the server has received it. From there, we can infer that sufficiently old
keys are stale. We can't remove stale keys immediately because some older
notices may still be in flight, but after a grace period they can go.

The timeout is set to 60 seconds, which is fairly high, but matches
Z_ReadWait's timeout.